made all drones available, drone gameobject not made in unity

+++ Assets/Scripts/Menu Scripts/SpeedSlider.cs
@@ -20,7 +20,7 @@ public class SpeedSlider : MonoBehaviour
     public void ValueChangeCheck()
     {
         Debug.Log(mySlider.value);
-        WorldProperties.GetSelectedDrone().setSpeed((int) mySlider.value);
+        WorldProperties.GetSelectedDrone().setSpeed(mySlider.value);
     }
 
 }
